Added server-side replication.
Extended IO control structure. Now command's (dnet_cmd) ID is an address
of the remote node to put given transaction to (it is used as before to
find the node which ID range covers command ID).

IO attribute now has two IDs: origin of the transaction and data
checksum. Origin (named origin) is used to find the history - it is ID
of the main object we write to. Checksum (named id) is used to specify
the ID of this particular transaction.

When transaction itself is written into the network, i.e. when it is not
tied to the main object (like file), origin and id may be the same.

When we write transaction which belongs to file, and store it somewhere
else, its history will only contain single ID which corresponds to the
ID of the file.

Example: let's suppose we want to write some data into the file named
'/tmp/cred' using single transformation function.

transform('/tmp/cred') = 1234
transform(content0)    = aaaa
transform(content1)    = bbbb

There will be 3 objects created with above IDs.
History for the first one will contain IDs of the all transactions ever
written into '/tmp/cred' file, namely aaaa and bbbb transactions.

History for each of the aaaa and bbbb transactions will only contain
single ID for their corresponding main object, namely 1234.

Using this technique we can determine if given transaction belongs to
some external object or is lost.