Added hash prefix (one byte) to implement virtual datacenters.
This feature allows to specify preconfigured prefix in every transaction
ID (the first byte). If new nodes are added with IDs starting from the
same number as in above hash function, then every transaction will go to
that ID.

Here is a virtual datacenter example.
Let's suppose we have to transformation functions setup on client:
dc1_sha1 and dc2_sha1. They will produce sha1 hash of the data with the
first byte set to 1 and 2 accordingly.

Now let's suppose we add two nodes with IDs being equal to 0x0100 and
0x0180 and then two nodes with 0x0200 and 0x0280 IDs. Now every
transaction made with above transformation functions will have its first
byte set either to 1 or 2, so there will be a guaranteed copy in the
first set of nodes and in the second set (sets are indexed by the first
ID byte, this set is a so called virtual datacenter).

Now we can implement multiple sets with different ID's first byte for
every datacenter we want to work with. If we only need to have two
copies, but there are more than 2 datacenters, machines from them can be
spread between two sets of nodes.