Added fcgi request completion flag.
When it is not set, fcgi request can not be used, usually because it
was already closed. This is a weak protection since data from one
request can be sent to another client, but it will only happen when
we exit early on timeout, which should be noticed in logs, and
timeouts changed appropriately.