library: made dnet_request_cmd() non-blocking
All its users (monitor, statistics, indexes, more?) expect it to be non-blocking.
It is not a bug to block, it continues to invoke callbacks asynchronously, but
calling thread was blocked. This is especially annoying when requesting large stats.

Also removed control structure dnet_request_cmd_priv, it was only used as a reference
counter for wrapper around wait mechanics to hold original callback and private data.