indexes: Fixed set_indexes in multibackend systems
There was an error when shard of indexes' list of key for this key is
stored at the machine but different backend as list of indexes for this
key. We have to send UPDATE_INTERNAL commands to other backend but
otherwise it was processed on the same one.

For fixing this issue there was implemented a mechanism which makes
possible to asynchronously send commands to other backends at the same
node and receive replies back to this backend's io pool.