net: Fixed race condition of route-list counter
In some cases it leaded to dropping results of route-list requests
during connection, i.e. this was possible if last route list reply
was received while last io event was in process. In that case last
route list replies would be dropped as route-list counter would be
equal to zero.